+++ src/main/java/io/xlorey/fluxloader/utils/Constants.java
@@ -36,6 +36,9 @@ public class Constants {
          */
         public static final ArrayList<String> WHITELIST_FLUXLOADER_FILES = new ArrayList<>() {{
                 add("io/");
+                add("imgui/");
+                add("com/google/");
+                add("org/yaml/");
         }};
 
         /**