+++ changelog.md
@@ -2,6 +2,9 @@
 - Added missing `Configuration` data types
 - Added a method to the `Plugin` class to get the config folder inside the folder with all plugins (`./plugins/{pluginName}/...`)
 - Added item IDs to `Wiki`
+- Added plugin metadata structure version tracking
+- Added a centralized registry for all plugin data
+- Changed basic widget templates
 - The event and patch system has been rewritten
 
 # FluxLoader v0.7.1